+++ src/app/logincheck/page.tsx
@@ -1,21 +1,6 @@
-"use client";
-import Loading from "@/components/common/Loading";
-import { useSession } from "next-auth/react";
-import { useRouter } from "next/navigation";
-import React, { useEffect } from "react";
+import LoginCheck from "@/components/common/LoginCheck";
+import React from "react";
 
 export default function page() {
-  const { data: session } = useSession();
-  const router = useRouter();
-  useEffect(() => {
-    if (
-      session &&
-      (session?.user.displayName === null || !session?.user.displayName)
-    ) {
-      router.replace("/signup");
-    } else {
-      router.replace("/");
-    }
-  }, [session]);
-  return <Loading />;
+  return <LoginCheck />;
 }
