+++ src/components/atoms/Card.tsx
@@ -71,6 +71,7 @@ const Card = ({ type, canBook = false, data }: CardProps<CardType>) => {
         <Modal
           description="로그인이 후 이용 가능합니다."
           buttonText="로그인"
+          isCancelBtn={true}
           onClick={signIn}
           setVisible={handleVisible}
         />
