+++ src/components/molecules/DisplayNameInput.tsx
@@ -46,7 +46,7 @@ export default function DisplayNameInput() {
 
     useEffect(() => {
       if (!session || session?.user.displayName) router.replace("/");
-    }, [session]);
+    }, [session, router]);
 
   return (
     <section className="w-full flex flex-col gap-10 items-center justify-center">
